+++ modules/icon_border_progress/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### v1.3.1
+
+<details><summary>See Changes</summary>
+
+- Use Bubble Card's automatic background colours when `background_color` is not set, instead of defaulting to
+  `var(--bubble-icon-background-color)`.
+  This fixes background colours based on entity states.
+
+</details>
+
 ### v1.3.0
 
 <details><summary>See Changes</summary>
@@ -7,13 +17,15 @@
 - Removed `start_angle` configuration option, as it no longer makes sense with varying sizes.
   This will be superseded by a new option to offset the starting point in the future.
 - Fixed border shape for buttons that are not always square/circular.
+
 </details>
 
 ### v1.2.0
 
 <details><summary>See Changes</summary>
 
-- Added automatic border radius matching based on CSS variables (`--bubble-icon-border-radius`, `--bubble-border-radius`)
+- Added automatic border radius matching based on CSS variables (`--bubble-icon-border-radius`,
+  `--bubble-border-radius`)
 - Added support for custom border radius per button using `border_radius` configuration option
 - Added support for custom start angle per button using `start_angle` configuration option